refactor(handlers): delete the deprecated intent.ts barrel (Stage E)
Stage B left intent.ts as a 38-line `@deprecated` back-compat barrel re-exporting
the handler surface from the split modules. This repoints every consumer at the
owning module and removes the barrel.

- 26 import sites across 16 files (src/cli.ts, src/server/server.ts, and 14 test
  files, static + dynamic `await import()`) now import directly from dispatch.ts
  (handleToolCall/hasRequiredScope/TOOL_SCOPES), feature-cache.ts (the cache
  accessors), object-types.ts, search.ts, write-helpers.ts, cds-hints.ts, and
  shared.ts (ToolResult) instead of from ./intent.js.
- Deletes src/handlers/intent.ts and tests/unit/handlers/barrel-surface.test.ts
  (the barrel's surface lock — its job is done now that nothing re-exports).
- Drops the now-dangling `src/handlers/intent.ts` budget key (the dangling-key
  guard added earlier would otherwise fail the ratchet — working as intended).

No behaviour change: every symbol resolves to the same implementation it did
through the barrel. 3,712 tests green (−3 = the deleted barrel-surface block),
A1 tool snapshots byte-identical, typecheck/lint/build/check:sizes pass.
