+++ src/cli.ts
@@ -23,7 +23,8 @@ import { config } from 'dotenv';
 import { AdtClient } from './adt/client.js';
 import type { AdtClientConfig } from './adt/config.js';
 import { buildArgs, type OutputMode } from './cli-args.js';
-import { handleToolCall, type ToolResult } from './handlers/intent.js';
+import { handleToolCall } from './handlers/dispatch.js';
+import type { ToolResult } from './handlers/shared.js';
 import { getToolDefinitions } from './handlers/tools.js';
 import { detectFilename, lintAbapSource } from './lint/lint.js';
 import { parseArgs, resolveConfig } from './server/config.js';
